requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
anthropic==0.39.0
pyyaml==6.0.2
python-dotenv==1.0.1
